+++ src/main/java/com/example/TucShopBackend/Services/ProductsService.java
@@ -4,11 +4,10 @@ import com.example.TucShopBackend.Commons.ApiResponse;
 import com.example.TucShopBackend.Commons.CustomConstants;
 
 import com.example.TucShopBackend.Commons.Status;
-import com.example.TucShopBackend.DTO.CategoryDTO;
 
 import com.example.TucShopBackend.DTO.ProductsDTO;
-import com.example.TucShopBackend.DTO.VariantsDTO;
 import com.example.TucShopBackend.DTO.UpdateStockDTO;
+import com.example.TucShopBackend.DTO.VariantsDTO;
 import com.example.TucShopBackend.Models.Category;
 import com.example.TucShopBackend.Models.Products;
 import com.example.TucShopBackend.Repositories.CategoryRepository;
@@ -18,7 +17,6 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.io.InputStreamResource;
 import org.springframework.core.io.Resource;
 import org.springframework.core.io.UrlResource;
-import org.springframework.data.repository.query.Param;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.MediaType;
@@ -32,11 +30,9 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.sql.Timestamp;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import java.util.Set;
 
 @Service
 public class    ProductsService {
@@ -220,9 +216,37 @@ public class    ProductsService {
         productsRepository.deleteAll();
         return new ApiResponse(Status.Status_Ok,CustomConstants.PROD_DELETE,null  );
     }
-    public ApiResponse deleteProductById(Long id){
-        productsRepository.deleteById(id);
-        return new ApiResponse(Status.Status_Ok, CustomConstants.PROD_DELETE, null);
+    public  ApiResponse deleteProductById(Long id) {
+
+       Optional<Products> products = productsRepository.findById(id);
+       if(products.isPresent()){
+           String imgPath = products.get().getImage();
+           String[] path = imgPath.split("/");
+           String path1="";
+           for(Integer i = 6; i < path.length ; i++){
+               if(i == path.length-1){
+                   path1 +=  path[i];
+               }
+               else if (i==7){
+
+                   path1+= "products//";
+               }
+               else{
+                   path1 +=  path[i]+"//";
+               }
+           }
+           if (deleteProductImage(path1)){
+
+              productsRepository.deleteById(id);
+
+              return new ApiResponse(Status.Status_Ok, CustomConstants.PROD_DELETE, null);
+
+          }
+          else{
+              return new ApiResponse(Status.Status_ERROR, CustomConstants.PRODIMAGE_ERROR, null);
+          }
+       }
+       return new ApiResponse(Status.Status_Ok, CustomConstants.PROD_DELETE, null);
     }
 
     public ApiResponse updateById(Long id , ProductsDTO productsDTO) {
@@ -395,4 +419,24 @@ public class    ProductsService {
     }
 
 
+    public Boolean deleteProductImage(String path) {
+        String filepath = serverFilePath+"serverFiles//"+path;
+        File f = new File(filepath);
+
+        try {
+
+            if(f.delete()) {
+                System.out.println(f.getName() + " is deleted!");
+
+            } else {
+                System.out.println("Delete operation is failed.");
+            }
+        }
+        catch(Exception e)
+        {
+            System.out.println("Failed to Delete image !!");
+        }
+
+        return  true;
+    }
 }
